Testing config
Updated junit to junit4.
unit tests go in src/test.
Unit test for BookBean.

+++ src/main/java/beans/BookBean.java
@@ -10,10 +10,10 @@ public class BookBean {
 
 	private String bid;
 	private String title;
-	private int price;
+	private double price;
 	private String category;
 
-	public BookBean(String bid, String title, int price, String category) {
+	public BookBean(String bid, String title, double price, String category) {
 		this.bid = bid;
 		this.title = title;
 		this.price = price;
@@ -36,11 +36,11 @@ public class BookBean {
 		this.title = title;
 	}
 
-	public int getPrice() {
+	public double getPrice() {
 		return price;
 	}
 
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
